requirements.txt
# WeatherItBetter - Python Dependencies
# ===========================================

# Core Dependencies (Required)
# ---------------------------------------------

# Google Generative AI SDK for Gemini
# Must be >= 1.50.0 for compatibility with google-adk
google-genai>=1.50.0

# Google Agent Development Kit (ADK) for Vertex AI Memory Bank
# Requires google-genai >= 1.50.0 and websockets >= 15.0.1
google-adk[vertexai]>=1.19.0

# Websockets (required by google-adk)
websockets>=15.0.1

# HTTP requests for weather API
requests>=2.31.0,<3.0.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0


# Google Calendar Integration (Required)
# ---------------------------------------------

# Google authentication libraries
google-auth>=2.23.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.1.1,<1.0.0

# Google API client library
google-api-python-client>=2.100.0,<3.0.0


# Development Dependencies (Optional)
# ---------------------------------------------
# Uncomment for development work

# pytest>=7.4.0
# black>=23.9.0
# flake8>=6.1.0
# mypy>=1.5.0
